Move class that refs VS Code into vscode/ folder
