+++ src/shared/vscode/interfaces.ts
@@ -1,11 +1,20 @@
 import * as child_process from "child_process";
 import { CompletionItem, CompletionItemProvider, DebugConfigurationProvider, DebugSession, DebugSessionCustomEvent, MarkdownString, RenameProvider, TextDocument, TreeDataProvider, TreeItem, Uri } from "vscode";
 import { AvailableSuggestion, FlutterOutline, Outline } from "../analysis_server_types";
-import { TestStatus, VersionStatus } from "../enums";
-import { DebugCommandHandler } from "../interfaces";
+import { FlutterService, FlutterServiceExtension, TestStatus, VersionStatus } from "../enums";
 import { EmittingLogger } from "../logging";
 import { WorkspaceContext } from "../workspace";
 import { Context } from "./workspace";
+
+export interface DebugCommandHandler {
+	flutterExtensions: {
+		serviceIsRegistered(service: FlutterService): boolean;
+		serviceExtensionIsLoaded(extension: FlutterServiceExtension): boolean;
+	};
+	handleDebugSessionStart(session: DebugSession): void;
+	handleDebugSessionEnd(session: DebugSession): void;
+	handleDebugSessionCustomEvent(e: DebugSessionCustomEvent): void;
+}
 
 export interface InternalExtensionApi {
 	analyzerCapabilities: {
